Add BBRUN_SCRIPT, remove '/lt' mode, fix shebangs, use ~/.local/tmp
Using a BBRUN_SCRIPT vairable instead of a mode allows the script
author to locate the original script file and location if needed.

Shebangs were also broken if not using `/usr/bin/env`, this is now
always used post-build.

I also moved the built script temp file to ~/.local/tmp so as not
to bleed users' personal scripts out to root system.